requirements.txt
streamlit>=1.22
pandas>=2.0
numpy>=1.25
scikit-learn>=1.3
xgboost>=1.7
shap>=0.42
plotly>=5.16
sentence-transformers>=2.2
faiss-cpu>=1.8
langchain>=0.1
openai>=0.28
python-dotenv
fastapi>=0.95
uvicorn[standard]
pytest
black
ruff
jinja2
gitpython
requests
google-generativeai>=0.6.0
httpx>=0.24.0
